pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pystar"
version = "0.1.0"
description = "Spatial Transcriptomics Analysis and Registration Pipeline"
authors = [
    {name = "Zhurui Zenglab"}
]
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Image Processing",
]
keywords = ["spatial-transcriptomics", "image-processing", "bioinformatics", "microscopy"]
dependencies = [
    "numpy",
    "pandas",
    "tifffile",
    "dask",
    "distributed",
    "scipy",
    "scikit-image",
    "matplotlib",
    "seaborn",
    "tqdm",
    "pyyaml",
    "pydantic",
    "opencv-python-headless",
]

[project.optional-dependencies]
gpu = [
    "cupy",
    "dask-cuda",
]
dev = [
    "pytest",
    "jupyterlab",
    "ipykernel",
]

[project.urls]
Homepage = "https://github.com/yourusername/pystar"
Repository = "https://github.com/yourusername/pystar"
Documentation = "https://github.com/yourusername/pystar#readme"

[tool.hatch.build.targets.wheel]
packages = ["pystar"]
